chore: Move form into parent component, make each form into a component and make the shared components folder  more structured
